Phase 1: decompose SymbolExtractor
Extract cohesive, stateless helpers from SymbolExtractor into focused modules:
- usr_decoder.py: USR -> human-readable display name conversion
- cursor_utils.py: get_qualified_name, extract_namespace
- documentation_extractor.py: brief/raw doc comment extraction
- signature_builder.py: human-readable C++ signature construction
- alias_extractor.py: typedef/using/type-alias-template extraction

SymbolExtractor now imports and delegates to these modules, shrinking
from ~1600 to ~1000 lines. Update call_graph_service and tests to import
from the new modules.

Validation: make format-check, lint, type-check pass; make test:
1432 passed, 17 skipped, 4 xpassed
