+++ tests/test_human_readable_signature.py
@@ -10,51 +10,55 @@ from unittest.mock import Mock
 import pytest
 
 from clang_index_mcp.cpp_analyzer import CppAnalyzer
-from clang_index_mcp._symbols.symbol_extractor import SymbolExtractor
+from clang_index_mcp._symbols.signature_builder import (
+    build_human_readable_signature,
+    extract_params_from_type_spelling,
+    extract_trailing_qualifiers,
+)
 
 
 class TestExtractParamsFromTypeSpelling:
     """Test the static helper that extracts params from C type notation."""
 
     def test_simple_params(self):
-        assert SymbolExtractor._extract_params_from_type_spelling("void (int, double)") == "int, double"
+        assert extract_params_from_type_spelling("void (int, double)") == "int, double"
 
     def test_no_params(self):
-        assert SymbolExtractor._extract_params_from_type_spelling("void ()") == ""
+        assert extract_params_from_type_spelling("void ()") == ""
 
     def test_single_param(self):
         assert (
-            SymbolExtractor._extract_params_from_type_spelling("int (const char *)") == "const char *"
+            extract_params_from_type_spelling("int (const char *)") == "const char *"
         )
 
     def test_reference_param(self):
-        result = SymbolExtractor._extract_params_from_type_spelling("void (const std::string &)")
+        result = extract_params_from_type_spelling("void (const std::string &)")
         assert result == "const std::string &"
 
     def test_function_pointer_param(self):
         """Nested parens from function pointer params should be handled."""
-        result = SymbolExtractor._extract_params_from_type_spelling("void (void (*)(int), int)")
+        result = extract_params_from_type_spelling("void (void (*)(int), int)")
         assert result == "void (*)(int), int"
 
     def test_empty_string(self):
-        assert SymbolExtractor._extract_params_from_type_spelling("") == ""
+        assert extract_params_from_type_spelling("") == ""
 
     def test_none_input(self):
-        assert SymbolExtractor._extract_params_from_type_spelling(None) == ""
+        assert extract_params_from_type_spelling(None) == ""
 
     def test_no_parens(self):
-        assert SymbolExtractor._extract_params_from_type_spelling("int") == ""
+        assert extract_params_from_type_spelling("int") == ""
 
     def test_template_params(self):
         """Template angle brackets in params should be preserved."""
-        result = SymbolExtractor._extract_params_from_type_spelling(
+        result = extract_params_from_type_spelling(
             "void (std::map<int, int>, std::vector<std::string>)"
         )
         assert result == "std::map<int, int>, std::vector<std::string>"
 
     def test_const_method_type(self):
         """Const qualifier after closing paren should not be included in params."""
-        result = SymbolExtractor._extract_params_from_type_spelling("void (int) const")
+        result = extract_params_from_type_spelling("void (int) const")
         assert result == "int"
 
 
@@ -62,27 +66,27 @@ class TestExtractTrailingQualifiers:
     """Test the static helper that extracts const/noexcept/etc. qualifiers."""
 
     def test_const_qualifier(self):
-        result = SymbolExtractor._extract_trailing_qualifiers("void (int) const")
+        result = extract_trailing_qualifiers("void (int) const")
         assert result.strip() == "const"
 
     def test_no_qualifier(self):
-        result = SymbolExtractor._extract_trailing_qualifiers("void (int)")
+        result = extract_trailing_qualifiers("void (int)")
         assert result == ""
 
     def test_const_noexcept(self):
-        result = SymbolExtractor._extract_trailing_qualifiers("void (int) const noexcept")
+        result = extract_trailing_qualifiers("void (int) const noexcept")
         assert "const" in result
         assert "noexcept" in result
 
     def test_empty_string(self):
-        assert SymbolExtractor._extract_trailing_qualifiers("") == ""
+        assert extract_trailing_qualifiers("") == ""
 
     def test_none_input(self):
-        assert SymbolExtractor._extract_trailing_qualifiers(None) == ""
+        assert extract_trailing_qualifiers(None) == ""
 
     def test_nested_parens_only_checks_outermost(self):
         """Qualifier extraction should use the last top-level close paren."""
-        result = SymbolExtractor._extract_trailing_qualifiers("void (void (*)(int)) const")
+        result = extract_trailing_qualifiers("void (void (*)(int)) const")
         assert result.strip() == "const"
 
 
@@ -128,7 +132,7 @@ class TestBuildHumanReadableSignature:
             result_type_spelling="void",
             args=[("int", "x"), ("const std::string &", "y")],
         )
-        result = analyzer.context.symbol_extractor._build_human_readable_signature(cursor)
+        result = build_human_readable_signature(cursor)
         assert result == "void processData(int x, const std::string & y)"
 
     def test_zero_arg_function(self, analyzer):
@@ -138,7 +142,7 @@ class TestBuildHumanReadableSignature:
             result_type_spelling="void",
             args=[],
         )
-        result = analyzer.context.symbol_extractor._build_human_readable_signature(cursor)
+        result = build_human_readable_signature(cursor)
         assert result == "void doSomething()"
 
     def test_const_method(self, analyzer):
@@ -148,7 +152,7 @@ class TestBuildHumanReadableSignature:
             result_type_spelling="int",
             args=[],
         )
-        result = analyzer.context.symbol_extractor._build_human_readable_signature(cursor)
+        result = build_human_readable_signature(cursor)
         assert result == "int getValue() const"
 
     def test_template_function_fallback(self, analyzer):
@@ -159,7 +163,7 @@ class TestBuildHumanReadableSignature:
             result_type_spelling="void",
             args=[],  # Templates often return empty args
         )
-        result = analyzer.context.symbol_extractor._build_human_readable_signature(cursor)
+        result = build_human_readable_signature(cursor)
         assert result == "void convert(int, double)"
 
     def test_function_pointer_param(self, analyzer):
@@ -170,7 +174,7 @@ class TestBuildHumanReadableSignature:
             result_type_spelling="void",
             args=[("void (*)(int)", "callback"), ("int", "priority")],
         )
-        result = analyzer.context.symbol_extractor._build_human_readable_signature(cursor)
+        result = build_human_readable_signature(cursor)
         assert result == "void registerCallback(void (*)(int) callback, int priority)"
 
     def test_unnamed_params(self, analyzer):
@@ -181,7 +185,7 @@ class TestBuildHumanReadableSignature:
             result_type_spelling="void",
             args=[("int", ""), ("double", "")],
         )
-        result = analyzer.context.symbol_extractor._build_human_readable_signature(cursor)
+        result = build_human_readable_signature(cursor)
         assert result == "void process(int, double)"
 
     def test_no_cursor_type(self, analyzer):
@@ -189,7 +193,7 @@ class TestBuildHumanReadableSignature:
         cursor = Mock()
         cursor.spelling = "testFunc"
         cursor.type = None
-        result = analyzer.context.symbol_extractor._build_human_readable_signature(cursor)
+        result = build_human_readable_signature(cursor)
         assert result == ""
 
     def test_reference_and_pointer_params(self, analyzer):
@@ -199,7 +203,7 @@ class TestBuildHumanReadableSignature:
             result_type_spelling="bool",
             args=[("int *", "data"), ("const std::string &", "name")],
         )
-        result = analyzer.context.symbol_extractor._build_human_readable_signature(cursor)
+        result = build_human_readable_signature(cursor)
         assert result == "bool transfer(int * data, const std::string & name)"
 
     def test_no_return_type(self, analyzer):
@@ -212,7 +216,7 @@ class TestBuildHumanReadableSignature:
         )
         # When result_type is empty string, we still get "void" prefix
         # because result_type_spelling returns "" which is falsy
-        result = analyzer.context.symbol_extractor._build_human_readable_signature(cursor)
+        result = build_human_readable_signature(cursor)
         assert result == "MyClass(int value)"
 
     def test_get_arguments_raises_exception(self, analyzer):
@@ -223,7 +227,7 @@ class TestBuildHumanReadableSignature:
         cursor.result_type.spelling = "void"
         cursor.get_arguments = Mock(side_effect=Exception("not available"))
 
-        result = analyzer.context.symbol_extractor._build_human_readable_signature(cursor)
+        result = build_human_readable_signature(cursor)
         assert result == "void testFunc(int, double)"
